feat: rework settings screen and improve workers

+++ app/src/main/java/io/github/turtlepaw/mindsky/components/post/DiscoveryContext.kt
@@ -19,7 +19,7 @@ enum class InsightType {
 
 @Composable
 fun PostInsightsContext(
-    value: Float,
+    value: Double,
     type: InsightType,
     modifier: Modifier = Modifier,
 ) {
